IGNITE-14360 Refactor FileLockHolder for reusage
